As per the code review comments
