+++ app/models/collection.client.ts
@@ -1,64 +1,31 @@
-// import { runQuery } from "./utils";
-
-// export const fetchCollectionById = async (
-//   collectionId: string,
-//   productsAmount: number
-// ) => {
-//   const getCollectionByIdQuery = `query getCollectionById($id: ID!) {
-//     collection(id: $id) {
-//       title
-//       products(first: ${productsAmount}) {
-//         edges {
-//             node {
-//                 id
-//                 title
-//                 description
-//                 variants(first:1) {
-//                     edges {
-//                         node {
-//                             title
-//                             id
-//                             priceV2 {
-//                                 amount
-//                                 currencyCode
-//                             }
-//                         }
-//                     }
-//                 }
-//             }
-//         }
-//       }
-//     }
-//   }
-//   `;
-
-//   const params = {
-//     query: getCollectionByIdQuery,
-//     variables: { id: collectionId },
-//   };
-
-//   const data = await runQuery(params);
-//   console.log(JSON.stringify(data, null, 4));
-// };
-
-// export const fetchHomeCollection = () =>
-//   fetchCollectionById("gid://shopify/Collection/216731549859", 10);
-
 import { runQuery } from "./utils";
 
-export const getCollections = async (productId: string) => {
-  const getProductByIdQuery = `query {
-    collections(first: 2) {
-      edges {
-        node {
-          id
-          products(first: 5) {
-            edges {
-              node {
+export const fetchCollectionById = async (
+  collectionId: string,
+  productsAmount: number
+) => {
+  const getCollectionByIdQuery = `query getCollectionById($id: ID!) {
+    collection(id: $id) {
+      title
+      products(first: ${productsAmount}) {
+        edges {
+            node {
                 id
-              }
+                title
+                description
+                variants(first:1) {
+                    edges {
+                        node {
+                            title
+                            id
+                            priceV2 {
+                                amount
+                                currencyCode
+                            }
+                        }
+                    }
+                }
             }
-          }
         }
       }
     }
@@ -66,11 +33,13 @@ export const getCollections = async (productId: string) => {
   `;
 
   const params = {
-    query: getProductByIdQuery,
-    variables: { id: productId },
+    query: getCollectionByIdQuery,
+    variables: { id: collectionId },
   };
 
-  const result = await runQuery(params);
-  console.log(JSON.stringify(result, null, 4));
-  return result.data.product;
+  const data = await runQuery(params);
+  console.log(JSON.stringify(data, null, 4));
 };
+
+export const fetchHomeCollection = () =>
+  fetchCollectionById("gid://shopify/Collection/216731549859", 10);
