+++ mocks/DUMMY_DATA.tsx
@@ -1,0 +1,212 @@
+export const perks = [
+  {
+    name: "Free returns",
+    imageUrl:
+      "https://tailwindui.com/img/ecommerce/icons/icon-returns-light.svg",
+    description:
+      "Not what you expected? Place it back in the parcel and attach the pre-paid postage stamp.",
+  },
+  {
+    name: "Same day delivery",
+    imageUrl:
+      "https://tailwindui.com/img/ecommerce/icons/icon-calendar-light.svg",
+    description:
+      "We offer a delivery service that has never been done before. Checkout today and receive your products within hours.",
+  },
+  {
+    name: "All year discount",
+    imageUrl:
+      "https://tailwindui.com/img/ecommerce/icons/icon-gift-card-light.svg",
+    description:
+      'Looking for a deal? You can use the code "ALLYEAR" at checkout and get money off all year round.',
+  },
+  {
+    name: "For the planet",
+    imageUrl:
+      "https://tailwindui.com/img/ecommerce/icons/icon-planet-light.svg",
+    description:
+      "We’ve pledged 1% of sales to the preservation and restoration of the natural environment.",
+  },
+];
+
+export const collections = [
+  {
+    name: "Clothes",
+    href: "#",
+    imageSrc:
+      "https://tailwindui.com/img/ecommerce-images/home-page-04-collection-01.jpg",
+    imageAlt: "Woman wearing a comfortable cotton t-shirt.",
+  },
+  {
+    name: "Drinking",
+    href: "#",
+    imageSrc:
+      "https://tailwindui.com/img/ecommerce-images/home-page-04-collection-02.jpg",
+    imageAlt: "Man wearing a comfortable and casual cotton t-shirt.",
+  },
+  {
+    name: "Home & Living",
+    href: "#",
+    imageSrc:
+      "https://tailwindui.com/img/ecommerce-images/home-page-04-collection-03.jpg",
+    imageAlt:
+      "Person sitting at a wooden desk with paper note organizer, pencil and tablet.",
+  },
+
+  // TODO Implementing UI element to navigate between collections
+  // {
+  //   name: "Phone",
+  //   href: "#",
+  //   imageSrc:
+  //     "https://tailwindui.com/img/ecommerce-images/home-page-04-collection-01.jpg",
+  //   imageAlt:
+  //     "Person sitting at a wooden desk with paper note organizer, pencil and tablet.",
+  // },
+  // {
+  //   name: "Beach",
+  //   href: "#",
+  //   imageSrc:
+  //     "https://tailwindui.com/img/ecommerce-images/home-page-04-collection-02.jpg",
+  //   imageAlt:
+  //     "Person sitting at a wooden desk with paper note organizer, pencil and tablet.",
+  // },
+  // {
+  //   name: "Bags",
+  //   href: "#",
+  //   imageSrc:
+  //     "https://tailwindui.com/img/ecommerce-images/home-page-04-collection-02.jpg",
+  //   imageAlt:
+  //     "Person sitting at a wooden desk with paper note organizer, pencil and tablet.",
+  // },
+  // {
+  //   name: "Covid-19",
+  //   href: "#",
+  //   imageSrc:
+  //     "https://tailwindui.com/img/ecommerce-images/home-page-04-collection-03.jpg",
+  //   imageAlt:
+  //     "Person sitting at a wooden desk with paper note organizer, pencil and tablet.",
+  // },
+];
+
+export const trendingProducts = [
+  {
+    id: 1,
+    name: "Leather Long Wallet",
+    color: "Natural",
+    price: "$75",
+    href: "#",
+    imageSrc:
+      "https://tailwindui.com/img/ecommerce-images/home-page-04-trending-product-02.jpg",
+    imageAlt: "Hand stitched, orange leather long wallet.",
+  },
+  // More products...
+];
+
+export const footerNavigation = {
+  products: [
+    { name: "Bags", href: "#" },
+    { name: "Tees", href: "#" },
+    { name: "Objects", href: "#" },
+    { name: "Home Goods", href: "#" },
+    { name: "Accessories", href: "#" },
+  ],
+  company: [
+    { name: "Who we are", href: "#" },
+    { name: "Sustainability", href: "#" },
+    { name: "Press", href: "#" },
+    { name: "Careers", href: "#" },
+    { name: "Terms & Conditions", href: "#" },
+    { name: "Privacy", href: "#" },
+  ],
+  customerService: [
+    { name: "Contact", href: "#" },
+    { name: "Shipping", href: "#" },
+    { name: "Returns", href: "#" },
+    { name: "Warranty", href: "#" },
+    { name: "Secure Payments", href: "#" },
+    { name: "FAQ", href: "#" },
+    { name: "Find a store", href: "#" },
+  ],
+};
+
+export const navigation = {
+  categories: [
+    {
+      name: "Women",
+      featured: [
+        {
+          name: "New Arrivals",
+          href: "#",
+          imageSrc:
+            "https://tailwindui.com/img/ecommerce-images/mega-menu-category-01.jpg",
+          imageAlt:
+            "Models sitting back to back, wearing Basic Tee in black and bone.",
+        },
+        {
+          name: "Basic Tees",
+          href: "#",
+          imageSrc:
+            "https://tailwindui.com/img/ecommerce-images/mega-menu-category-02.jpg",
+          imageAlt:
+            "Close up of Basic Tee fall bundle with off-white, ochre, olive, and black tees.",
+        },
+        {
+          name: "Accessories",
+          href: "#",
+          imageSrc:
+            "https://tailwindui.com/img/ecommerce-images/mega-menu-category-03.jpg",
+          imageAlt:
+            "Model wearing minimalist watch with black wristband and white watch face.",
+        },
+        {
+          name: "Carry",
+          href: "#",
+          imageSrc:
+            "https://tailwindui.com/img/ecommerce-images/mega-menu-category-04.jpg",
+          imageAlt:
+            "Model opening tan leather long wallet with credit card pockets and cash pouch.",
+        },
+      ],
+    },
+    {
+      name: "Men",
+      featured: [
+        {
+          name: "New Arrivals",
+          href: "#",
+          imageSrc:
+            "https://tailwindui.com/img/ecommerce-images/mega-menu-01-men-category-01.jpg",
+          imageAlt:
+            "Hats and sweaters on wood shelves next to various colors of t-shirts on hangers.",
+        },
+        {
+          name: "Basic Tees",
+          href: "#",
+          imageSrc:
+            "https://tailwindui.com/img/ecommerce-images/mega-menu-01-men-category-02.jpg",
+          imageAlt: "Model wearing light heather gray t-shirt.",
+        },
+        {
+          name: "Accessories",
+          href: "#",
+          imageSrc:
+            "https://tailwindui.com/img/ecommerce-images/mega-menu-01-men-category-03.jpg",
+          imageAlt:
+            "Grey 6-panel baseball hat with black brim, black mountain graphic on front, and light heather gray body.",
+        },
+        {
+          name: "Carry",
+          href: "#",
+          imageSrc:
+            "https://tailwindui.com/img/ecommerce-images/mega-menu-01-men-category-04.jpg",
+          imageAlt:
+            "Model putting folded cash into slim card holder olive leather wallet with hand stitching.",
+        },
+      ],
+    },
+  ],
+  pages: [
+    { name: "Company", href: "#" },
+    { name: "Stores", href: "#" },
+  ],
+};
